+++ CHANGELOG.md
@@ -90,6 +90,7 @@ are welcomed to check the E-Mails generated for missed chat messages. @janopae
 - New function for deleting store wall posts via rest !1390 #9 @alex.simm
 - Get rid of /upload.php !1365 @\_fridtjof_
 - Removed unused delPost xhr function !1417 @alex.simm
+- Removed ActivityModel by moving functions to ActivityXhr !1434 #9 @alex.simm 
 
 ## Dev/Test/CI stuff
 - Add "linux" tag for finding CI servers !1332 @nicksellen